refactor(actionmenu): drop cjs; convert to esm only
File extensions on relative imports is an esm requirement:

https://nodejs.org/api/esm.html#esm_terminology

This will be nice if we ever get to the point of using native browser
imports, which I hope we do.

This is also required for the build:css process once it points at the
esm output.

These were initially added when tree shaking wasn't assumed. It allowed
importing only what is needed. Now tree shaking is the standard and
there is no CJS.

These were initially added when tree shaking wasn't assumed. It allowed
importing only what is needed. Now tree shaking is the standard and
there is no CJS.

BREAKING CHANGE: Remove file path imports (eg, 'packagename/react')
BREAKING CHANGE: Drop CJS build
